// types.ts
export type TransactionType = {
     id?: string;
     type: string;
     amount: number;
     category?: string;
     date: Date | Timestamp | string;
     description?: string;
     image?: any;
     uid?: string;
     walletId: string;
};
// ...
export type TransactionListType = {
     data: TransactionType[];
     title?: string;
     loading?: boolean;
     emptyListMessage?: string;
};

export type TransactionItemProps = {
     item: TransactionType;
     index: number;
     handleClick: Function;
};